alumni/requirements.txt
asgi-redis==1.4.3
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.15.1
channels==3.0.5
channels-redis==4.1.0
constantly==15.1.0
cryptography==41.0.3
daphne==3.0.2
Django==4.2.4
django-filter==23.3
djangorestframework==3.14.0
future==0.18.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Markdown==3.4.4
msgpack==1.0.5
msgpack-python==0.5.6
nodejs==0.1.1
npm==0.1.1
npx==0.1.1
numpy==1.25.2
optional-django==0.1.0
Pillow==10.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyOpenSSL==23.2.0
pytz==2023.3.post1
redis==5.0.0
serializers==0.2.4
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
zope.interface==6.0
